changed numberOfVehicles from float to Integer

+++ trafficcounter-backend/src/main/java/ch/bfh/trafficcounter/model/dto/geojson/VehicleAmountDto.java
@@ -19,7 +19,7 @@ public class VehicleAmountDto {
 	/**
 	 * Amount of vehicles
 	 */
-	private float numberOfVehicles;
+	private Integer numberOfVehicles;
 
 }
